meme-generator/src/requirements.txt
Flask==2.0.2
pandas==1.3.4
Pillow==8.4.0
python_docx==0.8.11
requests==2.26.0
